[Feature #163136709] refactor post a question endpoint
